+++ Waiter/SoupAsync.php
@@ -15,7 +15,7 @@ class SoupAsync extends WP_Async_Task {
 	/**
 	 * Prepare data for the asynchronous request
 	 *
-	 * @throws Exception If for any reason the request should not happen
+	 * @throws \Exception If for any reason the request should not happen
 	 *
 	 * @param array $data An array of data sent to the hook
 	 *